Add auth client credentials flow

+++ surl/app/schemas/auth.py
@@ -6,7 +6,7 @@ from pydantic import BaseModel
 class Token(BaseModel):
     access_token: str
     token_type: str
-    expires_in: int
+    expires_in: float
 
 
 class TokenPayload(BaseModel):
